+++ sncosmo/photdata.py
@@ -16,6 +16,7 @@ from .magsystems import get_magsystem
 # deprecated (private!) functions: make them available where they used to be
 from ._deprecated import standardize_data, normalize_data
 
+__all__ = ['select_data']
 
 PHOTDATA_ALIASES = OrderedDict([
     ('time', set(['time', 'date', 'jd', 'mjd', 'mjdobs', 'mjd_obs'])),
@@ -23,8 +24,11 @@ PHOTDATA_ALIASES = OrderedDict([
     ('flux', set(['flux', 'f'])),
     ('fluxerr', set(['fluxerr', 'fe', 'fluxerror', 'flux_error', 'flux_err'])),
     ('zp', set(['zp', 'zpt', 'zeropoint', 'zero_point'])),
-    ('zpsys', set(['zpsys', 'zpmagsys', 'magsys']))
+    ('zpsys', set(['zpsys', 'zpmagsys', 'magsys'])),
+    ('fluxcov', set(['cov', 'covar', 'covariance', 'covmat']))
     ])
+
+PHOTDATA_REQUIRED_ALIASES = ('time', 'band', 'flux', 'fluxerr', 'zp', 'zpsys')
 
 
 class PhotometricData(object):
@@ -34,6 +38,8 @@ class PhotometricData(object):
     ``zpsys``, which are all numpy arrays of the same length sorted by
     ``time``. This is intended for use within sncosmo; its implementation
     may change without warning in future versions.
+
+    Has attribute ``fluxcov`` which may be ``None``.
 
     Parameters
     ----------
@@ -53,7 +59,8 @@ class PhotometricData(object):
         else:
             raise ValueError('unrecognized data type')
 
-        mapping = alias_map(colnames, PHOTDATA_ALIASES)
+        mapping = alias_map(colnames, PHOTDATA_ALIASES,
+                            required=PHOTDATA_REQUIRED_ALIASES)
 
         self.time = np.asarray(data[mapping['time']])
         self.band = np.asarray(data[mapping['band']])
@@ -61,6 +68,8 @@ class PhotometricData(object):
         self.fluxerr = np.asarray(data[mapping['fluxerr']])
         self.zp = np.asarray(data[mapping['zp']])
         self.zpsys = np.asarray(data[mapping['zpsys']])
+        self.fluxcov = (np.asarray(data[mapping['fluxcov']])
+                        if 'fluxcov' in mapping else None)
 
         # ensure columns are equal length
         if isinstance(data, dict):
@@ -68,7 +77,20 @@ class PhotometricData(object):
                     len(self.fluxerr) == len(self.zp) == len(self.zpsys)):
                 raise ValueError("unequal column lengths")
 
-        # Sort by time, if necessary.
+        # handle covariance if present
+        if self.fluxcov is not None:
+            # check shape OK
+            n = len(self.time)
+            if self.fluxcov.shape != (n, n):
+                raise ValueError(
+                    "Flux covariance must be shape (N, N). Did you slice "
+                    "the data? Use ``sncosmo.select_data(data, mask)`` in "
+                    "place of ``data[mask]`` to properly slice covariance.")
+
+        # ensure sorted by time
+        self._sort_by_time()
+
+    def _sort_by_time(self):
         if not np.all(np.ediff1d(self.time) >= 0.0):
             idx = np.argsort(self.time)
             self.time = self.time[idx]
@@ -77,18 +99,26 @@ class PhotometricData(object):
             self.fluxerr = self.fluxerr[idx]
             self.zp = self.zp[idx]
             self.zpsys = self.zpsys[idx]
+            self.fluxcov = (None if self.fluxcov is None else
+                            self.fluxcov[np.ix_(idx, idx)])
 
     def __len__(self):
         return len(self.time)
 
     def __getitem__(self, key):
-        newdata = copy.copy(self)  # avoid going through __init__
+        newdata = copy.copy(self)
         newdata.time = self.time[key]
         newdata.band = self.band[key]
         newdata.flux = self.flux[key]
         newdata.fluxerr = self.fluxerr[key]
         newdata.zp = self.zp[key]
         newdata.zpsys = self.zpsys[key]
+        newdata.fluxcov = (None if self.fluxcov is None else
+                           self.fluxcov[np.ix_(key, key)])
+
+        # ensure sorted by time (key could have caused reordering)
+        newdata._sort_by_time()
+
         return newdata
 
     def normalized(self, zp=25., zpsys='ab'):
@@ -117,6 +147,8 @@ class PhotometricData(object):
         newdata.fluxerr = factor * self.fluxerr
         newdata.zp = np.full(len(self), zp, dtype=np.float64)
         newdata.zpsys = np.full(len(self), zpsys, dtype=np.array(zpsys).dtype)
+        if newdata.fluxcov is not None:
+            newdata.fluxcov = factor * factor[:, None] * self.fluxcov
 
         return newdata
 
@@ -128,6 +160,74 @@ def photometric_data(data):
         return PhotometricData(data)
 
 
+def select_data(data, index):
+    """Convenience function for indexing photometric data with covariance.
+
+    This is like ``data[index]`` on an astropy Table, but handles
+    covariance columns correctly.
+
+    Parameters
+    ----------
+    data : `~astropy.table.Table`
+        Table of photometric data.
+    index : slice or array or int
+        Row selection to apply to table.
+
+    Returns
+    -------
+    `~astropy.table.Table`
+
+    Examples
+    --------
+
+    We have a small table of photometry with a covariance column and we
+    want to select some rows based on a mask:
+
+    >>> data = Table([[1., 2., 3.],
+    ...               ['a', 'b', 'c'],
+    ...               [[1.1, 1.2, 1.3],
+    ...                [2.1, 2.2, 2.3],
+    ...                [3.1, 3.2, 3.3]]],
+    ...               names=['time', 'x', 'cov'])
+    >>> mask = np.array([True, True, False])
+
+    Selecting directly on the table, the covariance column is not sliced
+    in each row: it has shape (2, 3) when it should be (2, 2):
+
+    >>> data[mask]
+    <Table length=2>
+      time   x    cov [3]
+    float64 str1  float64
+    ------- ---- ----------
+        1.0    a 1.1 .. 1.3
+        2.0    b 2.1 .. 2.3
+
+    Using ``select_data`` solves this:
+
+    >>> sncosmo.select_data(data, mask)
+    <Table length=2>
+      time   x    cov [2]
+    float64 str1  float64
+    ------- ---- ----------
+        1.0    a 1.1 .. 1.2
+        2.0    b 2.1 .. 2.2
+
+    """
+    mapping = alias_map(data.colnames,
+                        {'fluxcov': PHOTDATA_ALIASES['fluxcov']})
+    result = data[index]
+    if 'fluxcov' in mapping:
+        colname = mapping['fluxcov']
+        fluxcov = result[colname][:, index]
+
+        # replace_column method not available in astropy 1.0
+        i = result.index_column(colname)
+        del result[colname]
+        result.add_column(fluxcov, i)
+
+    return result
+
+
 # Generate docstring: table of aliases
 
 # Descriptions for docstring only.
@@ -137,7 +237,8 @@ _photdata_descriptions = {
     'flux': 'Flux of observation',
     'fluxerr': 'Gaussian uncertainty on flux',
     'zp': 'Zeropoint corresponding to flux',
-    'zpsys': 'Magnitude system for zeropoint'
+    'zpsys': 'Magnitude system for zeropoint',
+    'fluxcov': 'Covariance between observations (array; optional)'
     }
 
 _photdata_types = {
@@ -146,7 +247,8 @@ _photdata_types = {
     'flux': 'float',
     'fluxerr': 'float',
     'zp': 'float',
-    'zpsys': 'str'
+    'zpsys': 'str',
+    'fluxcov': 'ndarray'
     }
 
 lines = [